+++ MVC/Controllers/VehicleModelsController.cs
@@ -1,12 +1,16 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Net;
 using System.Threading.Tasks;
 using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.EntityFrameworkCore;
 using MVC.Models;
-using Service;
+using MVC.ViewModels;
 using Service.DAL;
+using Service.EFModels;
 using X.PagedList;
 
 namespace MVC.Controllers
@@ -22,13 +26,21 @@ namespace MVC.Controllers
             _mapper = mapper;
         }
 
-        public async Task<IActionResult> Index(ModelsViewModel model)
-        {
-            PopulateMakesSelectList();
-            ViewData["PageSize"] = new SelectList(new List<int> { 2, 5, 10, 25, 50, 100 });
-            var models = await _vehicleService.GetPagedModels(model.Filtering, model.Sorting, model.Paging);
-            model.Models = _mapper.Map<IPagedList<VehicleModel>, IPagedList<VehicleModelViewModel>>(models);
-            return View(model);
+        public async Task<IActionResult> Index(FilteringModel filtering, SortingModel sorting, PagingModel paging)
+        {            
+            ViewBag.Filter = filtering;
+            ViewBag.Sort = sorting;
+            ViewBag.Paging = paging;
+            ViewBag.PageSize = new SelectList(new List<int> { 2, 5, 10, 25, 50, 100 }, paging.PageSize);
+
+            await PopulateMakesSelectList(filtering.FilterById);
+            var models = await _vehicleService.GetPagedModels(filtering, sorting, paging);
+            if (models.TotalItemCount == 0)
+            {
+                Response.StatusCode = (int)HttpStatusCode.NotFound;
+            }
+            var viewModel = _mapper.Map<IPagedList<IVehicleModel>, IPagedList<VehicleModelViewModel>>(models);
+            return View(viewModel);            
         }
                 
         public async Task<IActionResult> Details(int? id)
@@ -38,20 +50,18 @@ namespace MVC.Controllers
                 return BadRequest();
             }
 
-            var vehicleModel = await _vehicleService.ModelRepository.GetOneAsync(v => v.Id == id, "Make");
-            
-            if (vehicleModel == null)
+            var vehicleModelDTO = await _vehicleService.GetModelByIdAsync(id.GetValueOrDefault());
+            if (vehicleModelDTO == null)
             {
                 return NotFound();
             }
-            var viewModel = _mapper.Map<VehicleModelViewModel>(vehicleModel);
+            var viewModel = _mapper.Map<VehicleModelViewModel>(vehicleModelDTO);
             return View(viewModel);
         }
-
         
-        public IActionResult Create()
+        public async Task<IActionResult> Create()
         {
-            PopulateMakesSelectList();
+            await PopulateMakesSelectList();
             return View();
         }
         
@@ -63,17 +73,18 @@ namespace MVC.Controllers
             {
                 if (ModelState.IsValid)
                 {
-                    var vehicleModel = _mapper.Map<VehicleModel>(viewModel);
-                    _vehicleService.ModelRepository.Insert(vehicleModel);
+                    var vehicleModelDTO = _mapper.Map<IVehicleModel>(viewModel);
+                    _vehicleService.InsertModel(vehicleModelDTO);
                     await _vehicleService.SaveAsync();
-                    return RedirectToAction(nameof(Index));
+                    return RedirectToAction("Index");
                 }
             }
-            catch (Exception)
+            catch (DbUpdateException)
             {
-                ModelState.AddModelError("", "Unable to save changes. Try again, and if the problem persists see your system administrator.");
-            }
-            PopulateMakesSelectList(viewModel.MakeId);
+                Response.StatusCode = (int)HttpStatusCode.BadRequest;
+                ModelState.AddModelError("", "Unable to save. Try again, and if the problem persists see your system administrator.");
+            }    
+            await PopulateMakesSelectList(viewModel.MakeId);
             return View(viewModel);
         }
 
@@ -84,14 +95,14 @@ namespace MVC.Controllers
                 return BadRequest();
             }
 
-            var vehicleModel = await _vehicleService.ModelRepository.GetByIdAsync(id);
-            if (vehicleModel == null)
+            var vehicleModelDTO = await _vehicleService.GetModelByIdAsync(id.GetValueOrDefault());
+            if (vehicleModelDTO == null)
             {
                 return NotFound();
             }
-            var viewModel = _mapper.Map<VehicleModelViewModel>(vehicleModel);
-            PopulateMakesSelectList(viewModel.MakeId);
-            return View(viewModel);
+            var viewModel = _mapper.Map<VehicleModelViewModel>(vehicleModelDTO);
+            await PopulateMakesSelectList(viewModel.MakeId);
+            return View(viewModel);            
         }
 
         [HttpPost]
@@ -102,17 +113,18 @@ namespace MVC.Controllers
             {
                 if (ModelState.IsValid)
                 {
-                    var vehicleModel = _mapper.Map<VehicleModel>(viewModel);
-                    _vehicleService.ModelRepository.Update(vehicleModel);
+                    var vehicleModelDTO = _mapper.Map<IVehicleModel>(viewModel);
+                    _vehicleService.UpdateModel(vehicleModelDTO);
                     await _vehicleService.SaveAsync();
-                    return RedirectToAction(nameof(Index));
+                    return RedirectToAction("Index");
                 }
             }
-            catch (Exception)
+            catch (DbUpdateException)
             {
+                Response.StatusCode = (int)HttpStatusCode.BadRequest;
                 ModelState.AddModelError("", "Unable to save changes. Try again, and if the problem persists, see your system administrator.");
             }
-            PopulateMakesSelectList(viewModel.MakeId);
+            await PopulateMakesSelectList(viewModel.MakeId);
             return View(viewModel);            
         }
 
@@ -124,15 +136,16 @@ namespace MVC.Controllers
             }
             if (saveChangesError.GetValueOrDefault())
             {
+                Response.StatusCode = (int)HttpStatusCode.Conflict;
                 ViewBag.ErrorMessage = "Delete failed. Try again, and if the problem persists see your system administrator.";
             }
-            var vehicleModel = await _vehicleService.ModelRepository.GetOneAsync(v => v.Id == id, "Make");
-            if (vehicleModel == null)
+            var vehicleModelDTO = await _vehicleService.GetModelByIdAsync(id.GetValueOrDefault());
+            if (vehicleModelDTO == null)
             {
                 return NotFound();
             }
-            var viewModel = _mapper.Map<VehicleModelViewModel>(vehicleModel);
-            return View(viewModel);
+            var viewModel = _mapper.Map<VehicleModelViewModel>(vehicleModelDTO);
+            return View(viewModel);            
         }
 
         [HttpPost]
@@ -141,22 +154,26 @@ namespace MVC.Controllers
         {
             try
             {
-                var vehicleModel = await _vehicleService.ModelRepository.GetByIdAsync(id);
-                _vehicleService.ModelRepository.DeleteAsync(id);
+                var vehicleModelDTO = await _vehicleService.GetModelByIdAsync(id);
+                if (vehicleModelDTO == null)
+                {
+                    return NotFound();
+                }
+                await _vehicleService.ModelRepository.DeleteAsync(id);
                 await _vehicleService.SaveAsync();
             }
-            catch (Exception)
+            catch (DbUpdateException)
             {
                 return RedirectToAction("Delete", new { id, saveChangesError = true });
             }
             return RedirectToAction("Index");
         }
 
-        private void PopulateMakesSelectList(object selectedVehicleMake = null)
+        private async Task PopulateMakesSelectList(object selectedVehicleMake = null)
         {
-            var makes = _vehicleService.GetVehicleMakesDropDownList().Result;
-            var items = _mapper.Map<IEnumerable<VehicleMakeDropListModel>>(makes);
-            ViewData["MakesSelectList"] = new SelectList(items, "Id", "Name", selectedVehicleMake);
+            var makesDTO = await _vehicleService.GetMakes();
+            var items = _mapper.Map<IEnumerable<VehicleMakeDropListModel>>(makesDTO);
+            ViewBag.MakesSelectList = new SelectList(items.OrderBy(i => i.Name), "Id", "Name", selectedVehicleMake);
         }
 
         protected override void Dispose(bool disposing)
